+++ backend/controllers/PersonajesController.php
@@ -4,8 +4,6 @@ namespace backend\controllers;
 
 use Yii;
 use yii\filters\AccessControl;
-
-use common\models\Personajes;
 use common\models\PersonajesSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
@@ -16,6 +14,8 @@ use yii\web\NotFoundHttpException;
 class PersonajesController extends Controller
 {
     use \common\utilities\Permisos;
+    use \common\traitrollers\PersonajesComun;
+
     /**
      * @inheritdoc
      */
@@ -48,39 +48,6 @@ class PersonajesController extends Controller
     }
 
     /**
-     * Displays a single Personajes model.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionView($id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
-    }
-
-    /**
-     * Updates an existing Personajes model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionUpdate($id)
-    {
-        $model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
-    }
-
-    /**
      * Deletes an existing Personajes model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param integer $id
@@ -92,21 +59,5 @@ class PersonajesController extends Controller
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
-    }
-
-    /**
-     * Finds the Personajes model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return Personajes the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
-    {
-        if (($model = Personajes::findOne($id)) !== null) {
-            return $model;
-        }
-
-        throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
